Optional rewrite rules
I hesitated to add optionality back in because the programmer has no actual way to *express* this right now, and in general we've been putting all such features on the back-burner.  But it's only a few lines of code, and I wanted to make sure that the current way of doing things was at least compatible with optionality.

The new solution handles optionality in a different place.  Before, a rule NOT applying (e.g. instead of replacing `a` with `X`, just leaving it as `a`) was treated as a sort of matchAnythingElse, it was on what we're now calling the "second branch".  Now, however, it's on the same "branch" as a replacement.  The only difference is in the nucleus of the pattern match, we replace

    Cor($i:a, $o:X)

with

    Cor($i:a, $o:X) | Match($i:a, $i, $o)

The match doesn't have to be wrapped in a CorrespondExpr because matches always have the same length anyway, they don't need a CorrespondExpr to enforce that.