Introduced BoundingSet in the GrammarConvenience functions.
Our goal is to eliminate reliance on minReps and maxReps in Replace tests, i.e.
in testGrammarReplace. This is because Pat's new rewrite rules implementation
doesn't handle minReps and maxReps.

BoundingSet produces a query grammar that forces the from pattern to be present
in results (i.e. to effectively require minReps to be 1). This allows the test
writer to ignore testing uninteresting copy-through results. BoundingSet is called
as the query argument in the tests using it.

Queries can now be a Grammars or strings in addition to a StringDicts and
StringDict arrays.

testGrammarReplace has been partially updated to avoid using non-zero minReps and
non-infinite maxReps. All tests up to test 33j have been refactored. Some tests
make use of BoundingSet to reduce the number of results. I am committing the
code now to allow Pat to merge it into his branch to test them with the new
RewriteExpr.

test 33c in testGrammarReplace is currently failing. I don't understand why
removing minReps and maxReps for this test results in copy-through results since
the test does an inserting of a character with both beginsWith and endsWith
constraints. tests 33a and 33b do the same each with one of beginsWith or
endsWith, and these tests behave as before, each doing exactly one insertion, at
the start or the end of the input string, with no copy-through results.