Completed refactoring testGrammarReplace to eliminate use of minReps/maxReps.
I have figured out that the outputs produced by test 33c (now labelled 23c) are,
in fact, consistent with tests 33a (now 23a) and 33b (now 23b) and are what should
be expected, so I corrected the expected results for that test. I figured this out
while working on the corresponding deletion tests (34a-c, now labelled 24a-c). For
tests 23a and 23b, there are no copy-through results (for 0 insertions) because an
insertion can always be done at the start or end, whereas there are copy-through
results for test 23c, where the only insertion that can be done is when $i has
no content.

I removed unnecessary tests and renumbered those that were left.

A problem exists with tests 25b-c (formerly 35b-c). These tests die with a heap
out-of-memory error if a BoundingSet query is used. This is because we end up
going down a garden path on the BoundingSet side of the join and the limit for
$i on that side becomes DEFAULT_MAX_CHARS (100), and even for a 3 character
vocab, this generates a huge number of possibilities.

If, instead, we add a Count limiter on tape $i to the query, we no longer run out
of memory, but incorrect results are returned. Specifying a count of 3 on $i both
in the base grammar and in the query gives results allowing up to 6 characters on
$i. Using multiple Count grammars for the same tape clearly doesn't work as we
would expect.

I think we will probably have to change our counts for tests to be passed in to
the interpreter and added closer to the root.