Updated docs dependencies.
`npm install` in the docs folder failed for me. This is because Pull request #16
from dependabot on 2023-11-15 updated docusaurus to v3 without updating some of
the peer dependencies, such as react, react-dom, mdx-js/react, module-type-aliases.
This left the npm configuration in an un-installable state.

I think I have resolved all of the issues this caused. Specifically I updated the
following modules in package.json:
@docusaurus/core 3.0.0 -> 3.1.1
@docusaurus/preset-classic 3.0.0 -> 3.1.1
* @mdx-js/react 1.6.22 -> 3.0.1
clsx 1.2.1 -> 2.1.0
* prism-react-renderer 1.3.5->2.3.1
* react 17.0.2 -> 18.2.0
* react-dom 17.0.2 -> 18.2.0
* @docusaurus/module-type-aliases 2.4.1 ->3.1.1

I'm pretty sure that all of the modules marked with * in the above list required
the update (with perhaps prism-react-renderer not requiring updating). However,
doing all of the above brings them all up to date.

With these updates, two additional small changes to the source were required:
1. In docusaurus.config.js, the syntax of the require statement for
   prism-react-renderer themes changed.
2. The Highlight component needed to be defined as a custom component, and then
   imported in docs/tutorial/gs3.md where it is used.

The documentation now builds and looks the same as before.