A bit of clean up in the gramble docs.
Fixed the display a couple of tables that weren't displaying correctly.

Fixed the code in gs2.md to actually be valid Gramble code.

Sorted out the labelling of the first and second programs in the tutorial.
Also, removed a duplicate "third" program that was the same as the second (gs5.md).

The documentation in docs is very sparse. I think Pat's actual documenation is
elsewhere. We need to get it into docusaurus in the docs directory.