Fix to backwards-priority rewrites
(1) Usually rewrite rules have a fixed tape-priority from the ReplaceBlock structure.  The two tapes of the PreTapeExpr will always come in a particular order, and those get remapped through a RenameExpr to the "$i" and "$o" tapes that RewriteExpr expects, with the result that "$i" queries always end up coming first.

If we use the new RewriteExpr outside of that context, however, and "$o" queries come before "$i", we throw the "went nowhere" exception in the generator loop because there can end up being unsimplifiable content.  (Specifically, we delta out on "$o", and then the RewriteExpr ends up putting literal content on "$o" -- that, after all, is what Rewrites are *for* -- then that "$o" content won't ever be reached, because we'll never be calling an "$o" deriv again.)

While this isn't the sort of bug a programmer could encounter in a source-built grammar (because of the aforementioned ReplaceBlock structure), it was concerning, because it could be a symptom of something deeper.

Luckily, the fix ended up being easy.  RewriteExpr had been written so that any non-"$i" query is ignored (that is, it deltas to epsilon and derivs to null).  Changing this so that it accepts BOTH "$i" and "$o" queries works fine.

(2) In looking for the answer of this, we found a bug in how Outputs were collected/displayed.  It's minor and doesn't affect generation.  But takes a while to explain:

getOutputs() used to only be called at the very end of the generator loop, when it was time to yield outputs.  But when we moved from a global output collector to storing Outputs in the tree, we couldn't easily look at intermediate outputs for the purposes of (say) debug logs.  And we need that, that was our main signpost for where in the derivation we were.  So we started using getOutputs() for that, too.

getOutputs() was originally only implemented on Exprs that could survive until the end of that loop -- the tree of CursorExprs, ConcatExprs, and a few others, with the leaves of that tree all being TokenExprs.  The way getOutputs() has outputs at all is because TokenExpr.getOutputs() gives the string output of that token.  On the other hand, LiteralExpr.getOutputs() doesn't do this, those aren't handled yet and don't count as outputs.  (Backing up, we can see the whole generation process as turning a complex tree whose leaves are mostly LiteralExprs into a simpler tree whose leaves are mostly TokenExprs.)

But there was one place we used TokenExprs in the "not-yet-handled" tree: it's what MatchExpr creates on "$o".  (Why?  Well, why not?  TokenExpr is just a degenrate LiteralExpr with only one token, which lets us avoid thinking about indexing, LTR/RTL etc., and MatchExpr will only ever create one-token literals, so let's just use those.)  Buuuuut, it also meant that MatchExpr resulted in intermediate outputs that looked like they had spurious content on "$o".

This didn't affect generation, because eventually those intermediate outputs either succeed (in which case that "$o" token becomes actual output) or they fail and go away.  But it did mean that our "signposts" in debug logs were very confusing, because they seemed to say there existed "$o" outputs that appeared to be independent of the "$o" cursor that should be storing them.