Re-enabling epsilon-input rewrites
In the new rewrite rules, if the inputMaterial in a rewrite rule was just epsilon -- no pre, no input, no post, and not beginsWith or endsWith -- we'd get into an infinite recursion, as RewriteExpr.deriv would end up calling RewriteExpr.deriv on itself, infinitely.  (This is because the first branch of a RewriteExpr constructs a ConcatExpr, and if the first part of a concat is epsilon then it calls deriv on the second half -- which in the case of a RewriteExpr, is itself.)

This isn't inherent to the idea of a RewriteExpr, it was just the lazy way that I wrote the code for deriv.  (Since we got sampling working properly/efficiently, deriv() bodies have become difficult to write, and it was hard enough just getting the basics of RewriteExpr.deriv correct.)

What I had been doing was just writing the entire formula that would be created, and calling deriv on that.  That is, in pseudocode:

Deriv(Replace(abc->xyz)) =
    Deriv(
          (Cor(i:abc, o:xyz) · Replace(abc->xyz))
        | Join(Match(i->o, i:.) · Replace(abc->xyz), ~(i:abc))
    )

You can see here if "abc" is just epsilon, than this deriv ends up calling Deriv(Replace(abc->xyz)) again.  This is incorrect, and I knew it, but had bigger fish to fry at the moment.  Instead, I just forbade "abc" to be epsilon -- in constructReplaceBlocks, if we inferred "abc" to be potentially zero length, the rule was discarded and an error issued.

Really, the scope needs to be as follows, and this commit changes it to be so.

          (Deriv(Cor(i:abc, o:xyz)) · Replace(abc->xyz))
        | Deriv(Join(Match(i->o, i:.) · Replace(abc->xyz), ~(i:abc)))

This allows us to "turn on" epsilon-input rewrite rules again without fear of infinite recursion.

Doing this breaks a few tests, those that tested the permitted "beginsWith" epsilon rule, `eps -> X | #_`.  This was always allowed, but in a bit of a janky way (it relied on that deriv scope being wide, which in the bigger picture incorrect).

In the next commits, I'm going to try to do epsilon-input rewrite rules in a principled way, which (if I manange it) should also cover the special beginsWith case correctly.