New semantics for epsilon-input rewrite rules
As I've mentioned offhand several times, the new semantics of rules gives us the possibility of doing epsilon-input rewrite rules (henceforth EIRR) correctly.

One of the big questions of EIRR is how many times they can apply.  Epsilons, after all, don't increment the tape in any way, so they might cause infinite inserts between every pair of characters -- in fact, if optional==false, they MUST cause infinite inserts.  That's not what we actually want.

Instead, we want to do something like XFST rules of the form `[..] -> a` (if I remember the syntax correctly), which you can think of as requiring the incrementation of the tape in between invocations, and thus only doing a single unconditional insertion in between each pair of characters.  In XFST it's required that the programmer remember to use a [..] rule, but we want to do it automatically.

We struggled to think how to implement this in the old rule semantics.  There wasn't anything in the old semantics quite corresponding to the idea of incrementing one step; it was defined at a higher level than that.  But the new semantics is defined at the level of the individual step; the idea of [..] isn't all that different from the `Match(i->o, $i:.)`.

So what we do is something like what we did with RepeatExpr: add a delta term at the beginning.  (Unlike RepeatExpr, this isn't a repeat of the delta, because we DON'T want the possibility of infinite insertion here, it's just the delta.)

At first I only added it to [a variant of] branch 2: the didn't-match-the-pattern branch, where we match one dot and try again.  (Only difference is that the continuation is that of branch 1, because this IS a match.)

    delta(R(abc->xyz)) ⋅ Match(.) ⋅ branch1continuation

This MOSTLY worked but actually had some counterintuitive results in testGrammarReplaceEpsilon 23h and 23i, where you can either insert "a" or replace "h" with "a".  When we only add this delta term to branch 2, insertion is only possible when the next character to be matched is matched exactly, not when it's replaced.  (That is, it becomes impossible to insert and then subsequently replace.)  That might seem okay at first but looking at the results it causes odd gaps, where you say "Why can this form be generated and not that one?"

So we do have to take into account the possibility of either matching the pattern or matching the dot.  Simplifying the notation a bit it's:

    delta(R(abc->xyz)) ⋅ (Match(.) | Cor(abc->xyz)) ⋅ branch1continuation

Another thing to note is that this only handles the insertion PRECEDING each character, not the insertion at the end.  (E.g., it handles "ABC" going to "xAxBxC", but not "xAxBxCx".)  The insertion at the end we already handled correctly in a past commit -- if you remember, it needs to happen in the delta because at that point, $i is at end-of-string and will never be able to deriv.

This changes the outputs of several of the tests, because they assumed the possibility of inserting multiple times (e.g. "hh" going to "haah"), and now we're limited to one.