Rename source rule tests to "replace"
We used to make a distinction between "replaces" and "rules" (which were two different stages of replace rule development that we just needed different names for).  "Rule" is long gone in the codebase... except for in source tests, because it was tedious to rename all those files.  I finally got around to it.