"OI" tests for epsilon insertion
Since running new-style rewrite rules in "OI" order (where we evaluate the output tape before the input tape) proved troublesome, I also tried this out on the new epsilon-input rewrite rules.  I'm not surprised that they don't quite work.

Playing around a bit with different formulations of the EIRRs, it's basically due to the asymmetry of these rules, and the difficulty of coordinating what I-tape and O-tape queries have to do with them.  Allowing BOTH queries to proceed (which was the fix to the previous issue) gets us into a situation where both the I-tape and O-tape queries can result in the emission of the output material in the same generator "tick".  (E.g., if the rule is `eps -> X`, then the O-tape query will emit X, and then the I-tape query will also emit X (as the special epsilon case).  This would be okay if we allowed the repeated insertion of X into epsilon contexts, but we don't, we only want it inserted once.

Limiting the special epsilon case to I-tape query doesn't help here, because it's still only the I-tape query that's activating the special epsilon case.  For the O-tape query, that `X` is just ordinary matching, the pattern has "X" for the output tape and it's emitting that.

So I think this means that, ultimately, only I-tape queries can "drive" the replacement process.  When we get an O-tape query, we can't ALSO let it drive the replacement process; I think we'll have to do something else like let it return an "epsilon" transition.