Fix for OI-order epsilon-input replaces
As speculated in the last commit, having O-tape queries of ReplaceExprs being epsilon-token returns works fine, and gets us back to green on all tests.

I think there's a bit more work to do with making sure the up-tree structures can actually handle this epsilon-token return, though.  (I think this is the first time we've had the possibility of epsilons coming from the RIGHT branch of a join.)  But I wanted to commit this first.