Experimental new Correspond
I was exploring some i>o and o>i priority tests, which naturally go wrong when Correspond is involved, because it's one of our last remaining exprs that aren't priority-symmetric.

In doing so I came up with a new idea of how to do Correspond.  It doesn't quite work, but I wanted to commit it somewhere anyway.

The basic idea is that, instead of having a CorrespondExpr, we just construct an equivalent expression out of dots, repeats, joins, and a resurrected EpsilonLiteralExpr.  Basically there's a bounding set `(i:. ⋅ o:.)*` which we join to the whole thing, and then we also allow the emission of as many epsilons after the correspond-pattern on one (but not both) of the tapes.  (In other words, join any dict where I and O are the same length, to the correspond pattern except let one of them emit some epsilons to catch up.)

But this ends up not quite working in some corner cases, because you can end up in a situation where you're trying to delta out of the whole grammar, but can't because you still have to take another "epsilon step" in the replace rule.

I have some concern we may be able to replicate this issue in the "main" version of Correspond as well, if we set up the test in a certain way.