Bug in joining output tapes to rules
One thing we rarely (or maybe never) did before recently, was test replacement rules by joining grammars that specify the outputs.  We just joined them to grammars that specify the inputs.

Turns out it we do this, there's a corner case where we have a bug.  (It's the same bug as I just identified in the experimental branch newCorrespond, it didn't actually depend on those changes.)

The issue is when we have a rule with longer inputs than outputs (which is where Correspond comes into play by emitting epsilons as necessary), and it applies at the end of a word, AND we join it all to a grammar that specifies the output.  E.g. join (i:abcz, o:abX) with cz->X.  When we've consumed the `o:abX` and it's time to delta out of the o, we can't, because we're not done with the input tape yet, and that's what Correspond demands.  But we also can't emit an epsilon -- we're not in deriv at all here, and it's not going to be called because there's nothing on the other side of the join in the first place.  Delta isn't a place we can emit an epsilon, it's not that kind of function.

We may be able to handle this by just swapping the join order (so that rules are always on the left and thus always get a "dot" query), but it remains to be seen if that causes efficiency problems.