Fixed joining-a-replace-on-right bug
The secondary, minor bug found in the last commit ended up being a one-line fix, it was just that we weren't handling specific-token deriv queries correctly in PreTapeExpr.

When calling deriv w.r.t. the tape introduced by PreTapeExpr, we should have been constructing a dot query, like we do in CursorExpr.  (After all, the query on the new tape has no particular relationship to the original query on the other tape.)

But we didn't, we were actually doing a query-rename operation!  This worked fine when receiving dot-queries, which we do when PreTapeExpr is on the left side of joins.  When PreTapeExpr is on the RIGHT side of a join, however, the query is some particular token, and using that token to query on the new tape is obviously wrong.