Delta function as a generator
This was the start of an attempt to have Expr.delta return a generator of Deriv objects (just like Expr.deriv).  One of the things that makes it tricky to solve the bug found in the recent commits is that it depends on a situation in which ONE side of a join needs to delta out, and the other side needs to deriv along some epsilon transitions.  One solution to this is to make delta more like deriv -- that is, allow delta to "go" different places the same way deriv does.

(We've had this thought repeatedly over the last year, that some things that are a bit ugly would be simpler if delta/deriv had the same basic assumptions, or were the same function with different params.)

So I set about to do it, but in doing so realized that it was a bit larger project than anticipated.  Turning delta into deriv-like functions is easy, but rethinking everywhere we *call* delta is difficult.

I still think it'd be valuable, but it's a larger refactoring than I want to do, right before a 1.0 release.  I'll squash this bug a different way and then reconsider this later.