Abandoned attempt at end-rule joining fix
This is about halfway towards a fix for the bug where shortening rules don't apply at the ends of words when on the left side of a join.  This bug was originally discovered in this branch (but not exclusive to it).

The problem, again, is that it's possible to have the left side of a join delta-ing out, while the right side of a join still has some epsilon transitions to get to before it can delta out too.  So we either need to (a) add a case to JoinExpr.delta that handles this case, (b) add a case to JoinExpr.deriv that handles this, or (c) unify delta and deriv.

(c) would make the issue effectively go away but is a huge refactoring.  I attempted to do (a) in another branch, and it proved to be a bit bigger a job than I anticipated, because it completely changes what delta means and therefore we have to rethink every single call to delta.  So I'm attempting (b).

I thought it'd be easier to attempt (b) in this branch, because in this branch (but not main), epsilon transitions are actual expressions.  We could add a special case to JoinExpr.deriv when child1 is nullable, and send some sort of special epsilon or EOS query to the right side, that the EpsilonTokenExpr could be sensitive to.  (E.g., a deriv that says "Tell me where you can get to, if you can only travel along epsilons".)

But actually doing this here proves to be a bit of a mess, because the various expressions have to handle this special query in context-dependent ways.  Like in general, we don't want DotExpr matching epsilon queries -- dot means a character, and changing this breaks a lot of other things -- but epsilons count towards the sameLengthGuard of the new Correspond, so the DotExprs in there *should* match epsilon queries.  Or maybe we could handle these as `DotExpr(t) | EpsilonTokenExpr(t)`, but this EpsilonTokenExpr is again subtly different from its use on the other side of Correspond join.  (This echoes the issues in the last commit on this branch, that this same JoinExpr had to know whether it was part of a Correspond or not, to know how to handle epsilon transitions.)

These should all be raising alarm bells in the reader's head: the whole point of this branch was to explore whether Correspond could be expressed in terms of existing primitives instead, and now it seems that these aren't the familiar primitives, they differ in some minor but systematic ways in terms of the handling of epsilons.  So we may as well go back to the idea that Correspond is its own kind of expression, rather than introduce sublty-different behavior into existing primitives or introducing new versions of Dot, Join, etc. just to build Corresponds.

I still generally like this idea of Correspond using explicit EpsilonTokenExprs, but it's becoming clear that THIS particular implementation of Correspond is a dead end.