Fixing word-ending shortening rule join bug
This implements, roughly, the solution from the newCorrespond branch -- it didn't work there because it wasn't compatible with that expression of Correspond, but it works here.

The basic problem, going back to the very beginning, is:

  (a) If the output a replacement rule is shorter than the input (e.g., "ab -> c"), the output tape will basically "get ahead of" the input tape; unconstrained by the input it'll basically start us going down garden paths.  So we (like some other similar languages) end up needing replacement rule inputs and outputs to have the same length, and introduce "epsilon" transitions when necessary to make the output "longer" in terms of steps.

  (b) But the introduction of epsilon transitions gets us into a bind when it comes to the way we wrote delta and deriv.  There's a special case of join, where the left side is delta-able, but the right side still has some epsilon transitions to deriv, that isn't handled by either delta or deriv.  (There's a way to unify the two where this problem disappears, I think, but it's a huge refactoring and we don't have time right now.)

    (Backing up even further, Brzozowski's algorithm and its relatives are "epsilon free".  So in hindsight it's not a surprise that introducing epsilon transitions would mess up, e.g., the nature of delta.  But we didn't realize that when we wrote delta, necessitating either a big refactoring or a workaround.)

So we either have to add a special case to delta or a special case to deriv, because there is a valid path we aren't going down in either.  Adding it to delta proved also too much work, so I added it to deriv.

Now, if JoinExpr.deriv finds that child1 is nullable, it can send a special EpsilonTokenExpr deriv query to child2.  (Just like the usual deriv query is "Where can I get on tape T if I follow character C", this is "Where can I get to on tape T if I only follow epsilon transitions".)  This special query is rejected by just about every Expr, because there ISN'T any epsilon transition to follow on literals or dots or even epsilon.  But it's valid for Correspond.  (It happens that we don't actually have to DO anything in Correspond to accept it, it's just that Correspond is a rare Expr that can ignore the query and "go" elsewhere.  So it doesn't even need to check the query to see that it's an EpsilonTokenExpr, it's just doing its usual thing.)

Anyway, this solves this particular problem, but I don't love it, and it introduces a small performance penalty (about 10%), I think because it's introducing additional traversals of the tree in a common Expr.  So it'd be nice to find another solution later, but this is good enough for now.