Specifying OAuth scopes in sheets plugin
Our current plugin deploy script was overwriting the OAuth scopes on the deployed version.  It seems like the issue is that, in the absence of an explicit appsscript.json file, clasp creates a default one.  This default one doesn't specify the OAuth scopes needed.  Then this pushes to the project space and overwrites anything we've put there, resetting our OAuth requests to the default which is basically "your whole digital life".  (It asks for permission to overwrite, sure, but at the same time it's very tempting to say yes, and doing so undoes your work and it's not trivial to figure out the right answers again.)

NB: While in theory Google's able to read OAuth requirements right off our code, and does, they don't automatically use this information to populate the scope information.  Instead, I think the goal is to make us specify this manually, and then they read it off the code, too, to make sure we're telling the truth.

So I'm now creating an explicit appsscript.json file in the sheets_addon directory and the deploy script copies that to the deployment folder, and then clasp pushes it to the project space.