Refactored some grammar tests to follow best practices for calling testGrammar.
Specifically, we changed tests that called "describe" directly, and passed a
call to a local test function, to instead call testGrammar passing it a "desc"
parameter. In most cases we eliminated the need for a local test function.

For replace tests (in testGrammarSimple and testGrammarEpsilon) that specified
an io parameter that was expanded to a results parameter in a local testIO
function, we introduced an extension to testGrammar in testGrammarUtil, called
testGrammarIO, that does that conversion before calling testGrammar.

Various other tidying was also done to standardize the tests.