Adding intersection/join to regexes
We realized in talking with a new programmer that

    (a) people wanted single character negation in rule contexts (e.g., "change this to this when followed by a non-velar sound")
    (b) we had no way to express this

This is a different situation than "starts with not-a-velar", where we can change the scope to be what the programmer intends ("doesn't start with a velar"); if there's a way to do that for rule contexts I'm not seeing it.

But we could just express this in the regex, right?  It's just the intersection of . and the negated clause.  But it turns out we don't allow that, there's no intersection implemented in regexes.

Now we do, and it was an easy addition.

I put in some tests of this in filters and replaces.

I also put in tests of negated contexts WITHOUT this kind of . guard (e.g. just `_~(b|d)`) and I don't like the answers.  But I'm not sure what the answers SHOULD be, I just feel these shouldn't be it.