Allowing hide headers with empty values
If you had a hide header with an empty cell beneath it, it would try to hide a tape named "".  This, of course, isn't ever going to be a valid tape name and thus it'll never be found.

If a hide header with empty content sounds like an odd thing to try to do, consider a case where we're collecting a bunch of different stems from different sources, and they have different tapes (because there were different classes and other features necessary to get each sort of stem correct), and we don't want the resulting alternation to have any of those special tapes, only the tapes they share.

all_verbs =, embed, hide
           , c1_verbs, pos/trans
           , c2_verbs, pos
           , c3_verbs,
           , c4_verbs, pos/class

This wouldn't work.  Even though there's nothing to hide about c3_verbs, this syntactically wasn't possible because putting an empty cell under a hide header was invalid, it meant "hide the tape ''" and as mentioned, that's impossible.  As a result, we would never generate c3_verbs from this because they would be thrown out!

So now it doesn't mean that, it just means "don't do anything".