Eliminating Short in negative clause of replaces
I had originally included a ShortExpr in the negative clause of the new-style replace rule.  (The "negative clause" here means the NegationExpr that ReplaceExpr creates as one of its possible paths, the one that says "You can go forward along any character so long as you DON'T eventually match this pattern".  I clarify this because we're also talking about negative rule contexts here, which are *programmer* specified negations.)

Why did I include a ShortExpr?  Well, we usually do that in not-contains-like contexts because we want to "fail fast" as soon as we find something that fulfills the negative condition... and dollars to doughnuts I added it here in an attempt at premature optimization.

But in investigating some of the incorrect behavior of negative replace contexts, Darlene discovered that even a non-negated context like `_(h|eps)#` fails.  Looking at the resulting negative clause in debug logs, the resulting negative clause on the pass-through path ends up just being ~(eps), not ~(h|eps) -- we're throwing out the `h` somehow.  Why?  Well, because this is ~(Short(h|eps)), and that's ~(eps).  Oops.  So no, this isn't an appropriate context for shorts... we just didn't happen to be testing any other contexts where the alternatives were proper prefixes of each other.

This may not be the complete solution to the problems Darlene found, but I think it's clear that this isn't an appropriate Short().