Tutorial sheets in the Add-on
Expanded the tutorial somewhat, but we're still in the first few lessons.

However, when working through the tutorial, the reader will be getting a lot of table-by-table explanation of things, in a way that's potentially difficult to copy/paste into the sheet for hands-on play.

So I've added a sub-menu to the Gramble menu that lets the user make a new sheet starting from the source code of one of the tutorial examples (e.g. go to Gramble -> Tutorial sheets -> Tutorial 1).  I think this will make using the tutorial a lot smoother, and also it gives additional examples to the programmer who dove right in without reading the tutorial.

This is still done manually, though -- there's nothing that makes sure the tutorial and these pages stay in sync, so we'll have to be careful to remember this.