Test for singletape alternation bug
I added a test to illustrate the singletape alternation bug that Haokun found.  The issue is pretty simple although the fix raises some questions.

Background: SingleTape is basically a sort of Rename, just one that doesn't care what exactly the child's tape is named.  It says "My child can only have one tape, and whatever that tape is, I'll rename it to my tapename."  This is a convenience feature for the programmer, especially for embeds.  The programmer can have a table like Vowels, and embed that in a regex like `{Vowels}`, without having to worry about what to name Vowels' single tape.  (After all, the programmer CAN'T name the tape the proper thing in the first place, it's `$i` internally.)  They can just name it `text` or whatever is convenient, and the singleTape handler pass will see this one tape and rename it to `$i`.

However, what if we want a rule context that looks like this:

    {Vowels}|n_

That seems like a completely reasonable thing to say -- give me a vowel or `n`.  But the referent of {Vowels} has one tape, let's say `text`, whereas `n` also has one tape but its name is `$T`.  So the alternation as a whole has TWO tapes, {text, $T}, and is thus invalid in this context.

I had thought, much earlier, that the way SingleTape works is a bit hacky and we should probably have had the singleTape handler pass actually recurse through the tree and put a rename more tightly around the affected places.  E.g., instead of wrapping the whole grammar in a rename, put a Rename around each leaf like so:

    Rename(text->$i,Embed(Vowels)) | Rename($T->$i,$T:n)

This example convinces me that we should do something like this.

This raises some questions about what we should be able to recurse through.  For example, we certainly want to allow something like this:

    {Vowels}n_

So we should be able to recurse through SequenceGrammar, at least.

On the other hand, we don't want to recurse the singleTape handler all the way into the embedded symbols.  Our hypothesized recursive singleTape pass handler would effectively flatten any n-tape grammar into a 1-tape grammar, and this isn't what anyone would expect.