Fixing the alternation-singletape bug
I should have checked the code more thoroughly before typing all that out last time -- what I said we needed to do was already done!  It turns out handleSingleTapes already worked that way.

Whence the bug, then?  It was just an erroneous error check in tape-calc!  When rewriting tape-calc I just didn't foresee this edge case and put in a check that said "If there's >1 tape in the child, throw an error."  I took this out, obviously, and everything worked.

But in the meantime I also made some changes.  Before this I hadn't thought too hard about what SingleTape means; I was testing for "Does it work?" rather than "What does it mean?" and so wasn't really testing its behavior thoroughly.  So:

    * I added tests for various operations like seqs, alts, repeats, joins inside a SingleTape env.

    * We treated Embeds and Lits/Dots differently.  Whatever tape Embeds had, we renamed that.  But we only renamed literals if they were on tape `$T`.  This was fine -- by the nature of the parser, SingleTapes should never encounter literals that aren't on that tape -- but it's actually a bit of a surprise when you're writing a grammar using convenience functions.  ("Why is this working so differently than the should-be-equivalent version with embeds?")  So now it will rename a literal no matter what the original tape was.

Anyway I'm glad we had this bug because it made me think harder about what SingleTape means.  It's actually one of our stranger Grammars because SingleTape(G), where G denotes a language L, cannot be expressed as a function of L.  It actually changes the interpretation of the operators inside G.  We'll need to keep this in mind for proptesting, because otherwise meaning-preserving transformations to G won't be valid to perform inside SingleTape(G)!