Fixes to single-sheet downloads
The single-sheet download capability hasn't been changed since we changed the assignment syntax!  (E.g. where `Verb:` became `Verb=`.)  Since assigning stuff to collections is basically what single-sheeting does, it wasn't managing to do anything it all!  So I've fixed that.

Also, we were assigning a "Default" symbol automatically, to refer to the main sheet, but this no longer works because of changes to how defaulting/All work.  So I changed this to an embed of <MainSheetName>.All.