Factoring out finished Cursor into own Expr
Performance issues in the parse direction are suggesting to me that we need to experiment with a slightly different formulation of Cursor -- e.g. one that greedily consumes as many characters as it can -- and that we might have multiple kinds in the grammar depending on the execution plan.  (We already kinda have this with PreTape in any case.)

But before I get to that, I wanted to do a refactoring we had long considered: rather than having CursorExpr have a boolean member .finished and change its behavior based on that, to instead separate out that behavior into a FinishedExpr.  That behavior's not going to differ between different kinds of cursors, so they can all just become FinishedExprs when they're finished.