First attempt at greedy cursors
Several programmers have a problem with inefficient text parsing (e.g. querying {text: ninapenda} and getting back {tense: pres, subj:1sg, root:pend}).

I think part of this issue is that the breadth-first tape "search" -- where we make a single query to EACH tape rather than keep making queries to the same tape -- requires the sampling processs to make decisions about the other tapes too early.  For example, consider "ninapenda" above -- we decide what the root is on the first "loop" through the tapes, but only confirm that once we get to "p".  If it happens that we're right about the loop, great.  If it happens we're wrong, we've got to back up all the way to the point we made that decision!

A greedy Cursor -- depth-first on the text tape -- wouldn't have this problem.  We don't necessarily want to do EVERY tape like this, but when we can know that the tape isn't going to behave pathologically in the depth-first case, it should be possible.  Queries entered by the user can't exhibit pathological behavior -- they're literals -- and so joins of them should also be non-pathological.

This commit introduces a GreedyCursor.  It's quite simple: unlike a normal cursor, GreedyCursorExpr never calls forward on its child, it just does deltas/derivs on its own tape until it becomes a FinishedExpr, in which case it finally does call forward on its child.  (That's all a FinishedExpr does, after all.)

This commit puts GreedyCursors on all finite tapes (which is a rough approximation of non-pathological tapes... I hope).  I'll probably refine this later.

There are still a few snags, like this is fundamentally incompatible with evaluating a Correspond in the unexpected order.  (The GreedyCursor says "I can't query on t2 until I'm finished with t1", and the Correspond says "You can't finish with t1 until you're finished with t2!")  I'll think about that later, though; for now I want to commit what I have.